rename embedder-bridge to embedding-bridge
