+++ embedding-bridge/.envrc
