+++ embedding-bridge/.gitattributes
