+++ embedding-bridge/.gitignore
