+++ embedding-bridge/gradle.lock
