+++ embedding-bridge/gradle/wrapper/gradle-wrapper.properties
