+++ embedding-bridge/gradlew.bat
