+++ embedding-bridge/src/main/kotlin/com/nixops/embeddingbridge/EmbeddingBridgeApplication.kt
@@ -1,4 +1,4 @@
-package com.nixops.embedderbridge
+package com.nixops.embeddingbridge
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
