+++ embedding-bridge/src/main/kotlin/com/nixops/embeddingbridge/mapper/Mapper.kt
@@ -1,4 +1,4 @@
-package com.nixops.embedderbridge.mapper
+package com.nixops.embeddingbridge.mapper
 
 import org.mapstruct.Mapper
