+++ embedding-bridge/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 # base config
-spring.application.name=embedder-bridge
-server.port=8040
+spring.application.name=embedding-bridge
+spring.main.web-application-type=none
 
 # app config
 embedding-bridge.scraper.base-url=http://localhost:8080
